Isolated abstract models outside of models.py
Fixes #83.
